shortened the method in the main activity and overrode the WeatherResponse.tostring method
