+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Coord.kt
@@ -1,0 +1,10 @@
+package edu.uc.hornunjb.weatherwear.dto
+
+import com.google.gson.annotations.SerializedName
+
+class Coord {
+    @SerializedName("lon")
+    var lon: Float = 0.toFloat()
+    @SerializedName("lat")
+    var lat: Float = 0.toFloat()
+}
