+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Weather.kt
@@ -25,10 +25,3 @@ class Weather {
 
 
 
-
-class Coord {
-    @SerializedName("lon")
-    var lon: Float = 0.toFloat()
-    @SerializedName("lat")
-    var lat: Float = 0.toFloat()
-}
